[bphh-1406] Refactor computed compliance setup modal into sub components
